#75 Moved Bezier and Spline implementation to their own files.
