+++ graphics/coord.h
@@ -297,99 +297,6 @@ inline TransformMatrix rotation(double degrees, Coord center)
            translation(center.x(), center.y());
 }
 
-// Coefficient matrix used to calculate a Bezier curve
-inline TransformMatrix bezier_matrix()
-{
-    return TransformMatrix(
-        { -1, +3, -3, +1 },
-        { +3, -6, +3,  0 },
-        { -3, +3,  0,  0 },
-        { +1,  0,  0,  0 }
-    );
-}
-
-// Vector used to calculate points in each step of a Bezier curve
-inline TransformVector bezier_vector(double step)
-{
-    return TransformVector::of_step(step) * bezier_matrix();
-}
-
-// Coefficient matrix used to calculate a Spline curve
-inline TransformMatrix spline_matrix()
-{
-    return TransformMatrix(
-        { -1.0/6.0,      0.5,    -0.5, 1.0/6.0 },
-        {      0.5,     -1.0,     0.5,     0.0 },
-        {     -0.5,      0.0,     0.5,     0.0 },
-        {  1.0/6.0,  4.0/6.0, 1.0/6.0,     0.0 }
-    );
-}
-
-// Matrix used to calculate initial deltas of forward differences
-inline TransformMatrix delta_matrix()
-{
-    return TransformMatrix(
-        { 1.0, 1.0, 1.0, 0.0 },
-        { 6.0, 2.0, 0.0, 0.0 },
-        { 6.0, 0.0, 0.0, 0.0 },
-        { 0.0, 0.0, 0.0, 1.0 }
-    );
-}
-
-// Vector with initial deltas of forward differences
-inline TransformVector delta_vector(const TransformVector &v, double step)
-{
-    const TransformVector sv = TransformVector::of_step(step);
-    return TransformVector(
-        {
-          sv[0] * v[0],
-          sv[1] * v[1],
-          sv[2] * v[2],
-          sv[3] * v[3]
-        }
-    ) * delta_matrix();
-}
-
-// Calculate next delta vector based on previous one.
-inline TransformVector next_delta(const TransformVector &d)
-{
-    return TransformVector(
-        {
-            d[0] + d[1],
-            d[1] + d[2],
-            d[2],
-            d[3]
-        }
-    );
-}
-
-// Generate vertices using forward-differences technique.
-inline void generate_fd_vertices(
-    list<Coord> &vertices,
-    const TransformVector &vx,
-    const TransformVector &vy,
-    const TransformMatrix &m)
-{
-    constexpr double step = 0.025;
-
-    TransformVector dx = delta_vector(vx * m, step);
-    TransformVector dy = delta_vector(vy * m, step);
-
-    Coord current(dx[3], dy[3]);
-    vertices.push_back(current);
-
-    for (double t = 0.0; t <= 1; t += step)
-    {
-        const Coord next = current.translated(dx[0], dy[0]);
-        vertices.push_back(next);
-
-        dx = next_delta(dx);
-        dy = next_delta(dy);
-
-        current = next;
-    }
-}
-
 // Transform coord using transformation matrix.
 inline Coord operator * (const Coord &coord, TransformMatrix matrix)
 {
@@ -504,39 +411,3 @@ inline Coord at_step(double step, const Coord &start, const Coord &end)
     );
 }
 
-// Generate the vertices to represent a Bezier curve.
-inline list<Coord> bezier_vertices(const Coord &edge1, const Coord &control1, const Coord &control2, const Coord &edge2)
-{
-    const TransformVector vx = TransformVector::of_x(edge1, control1, control2, edge2);
-    const TransformVector vy = TransformVector::of_y(edge1, control1, control2, edge2);
-
-    list<Coord> coords;
-    for (double step = 0; step < 1 || equals(step, 1); step += 0.025)
-    {
-        const TransformVector b = bezier_vector(step);
-        coords.push_back(Coord(b * vx, b * vy));
-    }
-
-    return coords;
-}
-
-// Generate the vertices to represent a Spline curve.
-inline list<Coord> spline_vertices(vector<Coord> controls)
-{
-    constexpr size_t start = TransformVector::last_index;
-
-    assert(controls.size() > start);
-
-    list<Coord> result;
-
-    for (size_t i = start; i < controls.size(); i++)
-    {
-        generate_fd_vertices(
-            result,
-            TransformVector::of_x(controls, i),
-            TransformVector::of_y(controls, i),
-            spline_matrix());
-    }
-
-    return result;
-}
